get projects dynamically
