+++ utils/type.d.ts
@@ -1,0 +1,1 @@
+export type Project = 'morrent' | 'jobit';
